% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fortune_emptyR}
\alias{fortune_emptyR}
\title{Print a fortune from the 'fortunes' package}
\usage{
fortune_emptyR()
}
\value{
A list containing a fortune (an excerpt from an R forum or discussion).
}
\description{
Used to test detaching packages that other packages depend on. In this case, try
detaching 'fortunes' and then call this function, which needs fortunes.
}
\examples{
fortune_emptyR()

}
