% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_emptyR}
\alias{hello_emptyR}
\title{Return a confirmatory message}
\usage{
hello_emptyR()
}
\value{
Returns a character string.
}
\description{
Return a confirmatory message
}
\examples{
hello_emptyR()
#> [1] "emptyRpackage is installed!"
}
